+++ Game/src/mge/materials/TextureMaterial.cpp
@@ -94,7 +94,7 @@ void TextureMaterial::render(World* pWorld, Mesh* pMesh, const glm::mat4& pModel
 	glUniform1i(_steps, steps);
 
 	glUniform1i(_lightCount, pWorld->getLightCount());
-	for (size_t i = 0; i < pWorld->getLightCount(); i++)
+	for (int i = 0; i < pWorld->getLightCount(); i++)
 	{
 		_lightLocations[i].SetLight(&pWorld->getLightAt(i)->GetParams());
 	}
